create_gist {rCharts}	R Documentation

Function that takes a list of files and creates payload for API

Description

Function that takes a list of files and creates payload for API

Usage

  create_gist(filenames, description = "", public = TRUE)

Arguments

filenames
names of files to post

description
brief description of gist (optional)

public
whether gist is public (defaults to TRUE)

[Package rCharts version 0.4.5 Index]