renderChart {rCharts}	R Documentation

renderChart (use with Shiny)

Description

Use rCharts as Shiny output. First, use renderChart in server.R to assign the chart object to an Shiny output. Then create an chartOutput with the same name in #' ui.R. chartOutput is currently just an alias for htmlOutput.

Usage

  renderChart(expr, env = parent.frame(), quoted = FALSE)

Arguments

expr
An expression that returns a chart object

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

Author(s)

Thomas Reinholdsson, Ramnath Vaidyanathan

[Package rCharts version 0.4.5 Index]