showOutput {rCharts}	R Documentation

Use rCharts as Shiny output. First, use renderChart in server.R to assign the chart object to an Shiny output. Then create an chartOutput with the same name in #' ui.R. chartOutput is currently just an alias for htmlOutput.

Description

Use rCharts as Shiny output. First, use renderChart in server.R to assign the chart object to an Shiny output. Then create an chartOutput with the same name in #' ui.R. chartOutput is currently just an alias for htmlOutput.

Usage

  showOutput(outputId, lib = NULL, package = "rCharts")

Author(s)

Thomas Reinholdsson, Ramnath Vaidyanathan

[Package rCharts version 0.4.5 Index]