toChain {rCharts}	R Documentation

Converts an R list to a sequence of chained functions acting on a specified object.

Description

Converts an R list to a sequence of chained functions acting on a specified object.

Usage

  toChain(x, obj)

Author(s)

Ramnath Vaidyanathan

Examples

## Not run: 
toChain(list(showControls = TRUE, showDistX = TRUE), "chart")
## chart.showControls(true).showDistX(true)

## End(Not run)

[Package rCharts version 0.4.5 Index]