toJSONArray {rCharts}	R Documentation

Converts an R object to a JSON array of objects

Description

toJSON converts an R data frame to an object with key value pairs for each column. However, a lot of javascript charting libraries need an array of objects, one for each row of the data frame. This utility function does that.

Usage

  toJSONArray(obj, json = TRUE)

Author(s)

Ramnath Vaidyanathan

Examples

## Not run: 
toJSONArray(head(iris))

## End(Not run)

[Package rCharts version 0.4.5 Index]